Add files via upload
Basics of Artificial Intelligence is a Android Application I did and received a 100 on for ECPI while I was a student, this was my capstone project.

Known Errors are the highlight function which for some reason removes the HTML formatting in the application when the button is pressed to highlight a predetermined word or set of words, I do not know why this happens yet and have yet to find a way to fix this. My Instructor was unsure as well, I will look into this more to try to fix this bug.